20230125-avro_three_ways/python-avro/requirements.txt
avro
python-dateutil
